exit instead of reboot when running inside docker

+++ src/agent/onefuzz-agent/src/reboot.rs
@@ -90,11 +90,14 @@ impl Reboot {
 
     #[cfg(target_family = "unix")]
     pub fn invoke(&self) -> Result<()> {
-        info!("invoking local reboot command");
-
-        Command::new("reboot").arg("-f").status()?;
-
-        self.wait_for_reboot()
+        if std::path::Path.exists("/.dockerenv") {
+            info!("running inside docker, exiting instead of rebooting");
+            std::process::exit(0);
+        } else {
+            info!("invoking local reboot command");
+            Command::new("reboot").arg("-f").status()?;
+            self.wait_for_reboot()
+        }
     }
 
     #[cfg(target_family = "windows")]
